airflow/requirements.txt
pymongo==3.6.0
faker==18.13.0
psycopg2-binary==2.9.9